Obsługa turniejów

+++ EsportPlayerManager/Models/Tournament.cs
@@ -4,7 +4,7 @@ public class Tournament
 {
     public int Id { get; set; }
     public string Name { get; set; }
-    public decimal EntryFee { get; set; }
-    public decimal PrizePool { get; set; }
+    public int EntryFee { get; set; }
+    public int PrizePool { get; set; }
     public int MinSkillRequired { get; set; }
 }
